requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
configparser==7.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pillow_heif==0.21.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shutils==0.1.0
sniffio==1.3.1
soupsieve==2.7
squareup==42.0.0.20250416
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
